* vfy_increds.c: include int-proto.h for krb5_libdefault_boolean 	prototype.
	* t_ser.c (ser_eblock_test): ifdef out old eblock serialization
	test which is no longer called. (see 1999-09-01 ChangeLog)

	* t_kerb.c: Cast argument to fprintf to long to agree with format
	string.

	* t_deltat.c: If MIN is defined, undef before redefined as 60.

	* str_conv.c: Provide strptime prototype if the system header
	files fail to provide a prototype.

	* int-proto.h:  Add prototype for krb5_libdefault_boolean()

git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@12555 dc483132-0cff-0310-8789-dd5450dbe970

+++ src/lib/krb5/krb/t_kerb.c
@@ -178,7 +178,7 @@ main(argc, argv)
     retval = krb5_init_context(&ctx);
     if (retval) {
 	fprintf(stderr, "krb5_init_context returned error %ld\n",
-		retval);
+		(long) retval);
 	exit(1);
     }
     progname = argv[0];
